+++ src/renderer/components/Settings/FilterOptions.tsx
@@ -1,0 +1,44 @@
+
+import * as React from "react";
+
+import { ComboFilterSettings } from "@vinceau/slp-realtime";
+import { useDispatch, useSelector } from "react-redux";
+
+import { comboFilter } from "@/lib/realtime";
+import { Dispatch, iRootState } from "@/store";
+import { ComboForm } from "./ComboForm/ComboForm";
+import { ProfileSelector } from "./ComboForm/ProfileSelection";
+
+export const FilterOptions = () => {
+    const {currentProfile, comboProfiles} = useSelector((state: iRootState) => state.slippi);
+    // const [profile, setProfile] = React.useState<string>(currentProfile);
+    const profileOptions = Object.keys(comboProfiles);
+
+    let initial = comboFilter.getSettings();
+    const slippiSettings = comboProfiles[currentProfile];
+    if (slippiSettings) {
+        initial = comboFilter.updateSettings(JSON.parse(slippiSettings));
+    }
+    const dispatch = useDispatch<Dispatch>();
+    const onSubmit = (values: Partial<ComboFilterSettings>) => {
+        // console.log(values);
+        const valueString = JSON.stringify(values);
+        dispatch.slippi.saveProfile({
+            name: currentProfile,
+            settings: valueString,
+        });
+    };
+    const setProfile = (profile: string) => {
+        dispatch.slippi.setCurrentProfile(profile);
+    };
+    const onDelete = () => {
+        dispatch.slippi.deleteProfile(currentProfile);
+    };
+    return (
+        <div>
+            <h2>Filter Options</h2>
+            <ProfileSelector initialOptions={profileOptions} value={currentProfile} onChange={setProfile} />
+            <ComboForm initialValues={initial} onSubmit={onSubmit} onDelete={onDelete} />
+        </div>
+    );
+};
