+++ src/renderer/components/Settings/OBSSettings.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+export const OBSSettings = () => {
+    return (<div>
+        <h2>OBS Configuration</h2>
+    </div>);
+};
