+++ src/renderer/components/Settings/Settings.tsx
@@ -7,23 +7,19 @@ import {
     useRouteMatch,
 } from "react-router-dom";
 
-import { ComboFilterSettings } from "@vinceau/slp-realtime";
-import { useDispatch, useSelector } from "react-redux";
 import { Icon, Menu } from "semantic-ui-react";
 import styled, { css } from "styled-components";
 
-import { comboFilter } from "@/lib/realtime";
-import { Dispatch, iRootState } from "@/store";
 import { device } from "@/styles/device";
 import SlippiLogo from "@/styles/images/slippi-logo.svg";
 import OBSLogo from "@/styles/images/obs.svg";
 import { CustomIcon, Labelled } from "../Misc";
-import { SlippiPage } from "../Slippi";
+import { SlippiPage } from "./SlippiPage";
 import { ComboFinder } from "./ComboFinder";
-import { ComboForm } from "./ComboForm";
-import { ProfileSelector } from "./ProfileSelection";
 import { SoundSettings } from "./SoundSettings";
 import { TwitchIntegration } from "./TwitchIntegration";
+import { OBSSettings } from "./OBSSettings";
+import { FilterOptions } from "./FilterOptions";
 
 export const SettingsPage: React.FC<{
     showSettings: boolean;
@@ -169,45 +165,5 @@ export const SettingsPage: React.FC<{
                 </ContentColumn>
             </Wrapper>
         </SettingsContainer>
-    );
-};
-
-const OBSSettings = () => {
-    return (<div>
-        <h2>OBS Configuration</h2>
-    </div>);
-}
-
-const FilterOptions = () => {
-    const {currentProfile, comboProfiles} = useSelector((state: iRootState) => state.slippi);
-    // const [profile, setProfile] = React.useState<string>(currentProfile);
-    const profileOptions = Object.keys(comboProfiles);
-
-    let initial = comboFilter.getSettings();
-    const slippiSettings = comboProfiles[currentProfile];
-    if (slippiSettings) {
-        initial = comboFilter.updateSettings(JSON.parse(slippiSettings));
-    }
-    const dispatch = useDispatch<Dispatch>();
-    const onSubmit = (values: Partial<ComboFilterSettings>) => {
-        // console.log(values);
-        const valueString = JSON.stringify(values);
-        dispatch.slippi.saveProfile({
-            name: currentProfile,
-            settings: valueString,
-        });
-    };
-    const setProfile = (profile: string) => {
-        dispatch.slippi.setCurrentProfile(profile);
-    };
-    const onDelete = () => {
-        dispatch.slippi.deleteProfile(currentProfile);
-    };
-    return (
-        <div>
-            <h2>Filter Options</h2>
-            <ProfileSelector initialOptions={profileOptions} value={currentProfile} onChange={setProfile} />
-            <ComboForm initialValues={initial} onSubmit={onSubmit} onDelete={onDelete} />
-        </div>
     );
 };
